Align gstreamer build configuration with expermental's
Instead of selectively disabling gstreamer plugins, disable `auto_features`
and selectively enable them.

In this iteration match every plugin that is "auto" enabled in the previous
versions, pending an more in-depth look into `gst-plugins-bad` and `gst-plugins-ugly`

To minimize differences from experimental, also re-align the `.gitmodules` file
and move the ffmpeg submodule to the same name as in Valve's tree (`FFmpeg` -> `ffmpeg`).